chore(core): Move repository to state-management folder
