+++ packages/core/src/dispatcher.ts
@@ -4,7 +4,7 @@ import type { Document } from "./document"
 import { DoctypeUtils, IpfsApi, UnreachableCaseError } from '@ceramicnetwork/common';
 import DocID from "@ceramicnetwork/docid";
 import { DiagnosticsLogger, ServiceLogger } from "@ceramicnetwork/logger";
-import { Repository } from './repository';
+import { Repository } from './state-management/repository';
 import {
   buildQueryMessage,
   MsgType,
